pyproject.toml
[tool.poetry]
name = "storage"
version = "0.1.0"
description = ""
authors = ["aruzhant1209 <@aruzhant1209>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"
sqlalchemy = "^2.0.43"
asyncpg = "^0.30.0"
alembic = "^1.16.5"
python-dotenv = "^1.1.1"
dotenv = "^0.9.9"
psycopg2-binary = "^2.9.10"
pydantic-settings = "^2.10.1"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
